Fiddles with :define_method for Travis
Apparently you can't just use :define_method in `main:Object` in Ruby 1.9.3

```bash
undefined method `define_method' for main:Object (NoMethodError)
```

It needs to be explicitly called on the target class and, since it's also private, it requires the services of `send`. I don't think I'm going to leave this in here in the long run, but I'd like to play around with it a bit more to understand the nuances of `define_method` in different versions of Ruby.