Adds simplecov and coveralls to gemspec
This is the first time I've built a tested gem I intend to publish without using jeweler (which uses a rake task to generate the gemspec from a Gemfile). With nothing in the Gemfile but the `gemspec` line, I need to add these as development dependencies. I'll have to reread why you it's preferred this way, but it seems like the gems above don't belong in the gemspec, but in a test block in the Gemfile. The twitter gem does it this way. Still trying to understand if there are tradeoffs to either approach, or whether some of this is just convention.