Adds meta attributes to list responses
When a list of objects is returned, `total_count` and `next_href` attributes are returned with the response to support pagination for resources that are too large for a single query or have been explicitly partitioned with a limit parameter. These attributes are dynamically added to the returned Array object so they can be easily accessed.